ajout de printf dans libft

+++ ft_putnbr.c
@@ -13,7 +13,7 @@
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long n)
 {
 	if (n == -2147483648)
 		ft_putstr("-2147483648");
